Refactored display pod images method
