+++ src/main.rs
@@ -2,7 +2,8 @@ use anyhow::Result;
 use colored::*;
 use kelper::{
     cli::{Args, Commands, GetResource},
-    k8s::{display_pod_images, K8sClient},
+    k8s::K8sClient,
+    utils::display_pod_images,
 };
 
 #[tokio::main]
